Update x64 all ok

+++ Headers/KProtocol.h
@@ -778,7 +778,7 @@ struct tagNewDelRoleResponse : public tagDBSelPlayer
 // 2003.05.11
 struct tagDBSyncPlayerInfo : public tagProtoHeader
 {
-	size_t	dataLength;
+	unsigned int	dataLength;
 	char	szData[0];
 };
 
@@ -806,7 +806,7 @@ struct tagGuidableInfo : public tagProtoHeader
 	GUID guid;
 	WORD nExtPoint;			//可用的附送点
 	WORD nChangePoint;		//变化的附送点
-	size_t	datalength;
+	unsigned int	datalength;
 	char	szData[0];
 
 };
@@ -1047,7 +1047,7 @@ struct tagRegisterFamily : public tagProtoHeader
  */
 struct tagGS2GWSaveRole : public tagProtoHeader
 {
-	size_t	datalength;
+	unsigned int	datalength;
 	BYTE	szData[0];
 };
 
@@ -1079,7 +1079,7 @@ typedef struct
 struct TProcessData
 {
 	unsigned char	nProtoId;
-	size_t			nDataLen;//TRoleNetMsg时表示该Block的实际数据长度,TProcessData时表示Stream的实际数据长度
+	unsigned int	nDataLen;//TRoleNetMsg时表示该Block的实际数据长度,TProcessData时表示Stream的实际数据长度
 	unsigned long	ulIdentity;
 	bool			bLeave;
 	char			pDataBuffer[1];//实际的数据
